@@ -11,41 +11,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.0.0",
-    "@angular/common": "~8.0.0",
-    "@angular/compiler": "~8.0.0",
-    "@angular/core": "~8.0.0",
-    "@angular/forms": "~8.0.0",
-    "@angular/platform-browser": "~8.0.0",
-    "@angular/platform-browser-dynamic": "~8.0.0",
-    "@angular/router": "~8.0.0",
+    "@angular/animations": "~16.2.2",
+    "@angular/common": "~16.2.2",
+    "@angular/compiler": "~16.2.2",
+    "@angular/core": "~16.2.2",
+    "@angular/forms": "~16.2.2",
+    "@angular/material": "^16.2.2",
+    "@angular/platform-browser": "~16.2.2",
+    "@angular/platform-browser-dynamic": "~16.2.2",
+    "@angular/router": "~16.2.2",
+    "@ngtools/webpack": "^16.2.0",
     "angular-popper": "^2.0.1",
-    "bootstrap": "^4.3.1",
-    "jquery": "^3.4.1",
+    "bootstrap": "^5.3.1",
+    "jquery": "^3.7.0",
     "popper.js": "^1.15.0",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "rxjs-compat": "^6.6.7",
+    "tslib": "^2.6.2",
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "~8.0.0",
-    "@angular/compiler-cli": "~8.0.0",
-    "@angular/language-service": "~8.0.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "@angular-devkit/build-angular": "~16.2.0",
+    "@angular/cli": "~16.2.0",
+    "@angular/compiler-cli": "~16.2.2",
+    "@angular/language-service": "~16.2.2",
+    "@types/jasmine": "~4.3.5",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "~20.5.6",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~5.1.1",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.2",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "ts-node": "~10.9.1",
+    "tslint": "~5.20.1",
+    "typescript": "^4.9.3"
   }
 }
