@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import {GuardcomponentService} from './Guard/guardcomponent.service';
 import {ListCoursesComponent} from './list-courses/list-courses.component';
 import {EditCoursesComponent} from './edit-courses/edit-courses.component';
 import {OrganisationDetailsComponent} from './organisation-details/organisation-details.component';
@@ -7,7 +8,7 @@ import {AddCourseComponent} from './add-course/add-course.component';
 
 const routes: Routes =[
   {path:'',pathMatch:"full",redirectTo:"/organisationDetails"},
-  {path:'listCourses',component:ListCoursesComponent},
+  {path:'listCourses',component:ListCoursesComponent,canActivate:[GuardcomponentService]},
   {path:'editCourse',component:EditCoursesComponent},
   {path:'organisationDetails',component:OrganisationDetailsComponent},
   {path:'addCourse',component:AddCourseComponent}
